refactor: simplify data service
